+++ test/integrationtest/projecttest/graphql/generated.go
@@ -376,6 +376,13 @@ type ComplexityRoot struct {
 		Code   func(childComplexity int) int
 	}
 
+	CommerceCartQtyRestrictionResult struct {
+		IsRestricted        func(childComplexity int) int
+		MaxAllowed          func(childComplexity int) int
+		RemainingDifference func(childComplexity int) int
+		RestrictorName      func(childComplexity int) int
+	}
+
 	CommerceCartSelectedPaymentResult struct {
 		Processed      func(childComplexity int) int
 		ValidationInfo func(childComplexity int) int
@@ -780,6 +787,7 @@ type ComplexityRoot struct {
 
 	Query struct {
 		CommerceCart                     func(childComplexity int) int
+		CommerceCartQtyRestriction       func(childComplexity int, marketplaceCode string, variantCode *string, deliveryCode string) int
 		CommerceCartValidator            func(childComplexity int) int
 		CommerceCategory                 func(childComplexity int, categoryCode string, categorySearchRequest *dto1.CommerceSearchRequest) int
 		CommerceCategoryTree             func(childComplexity int, activeCategoryCode string) int
@@ -820,6 +828,7 @@ type QueryResolver interface {
 	CommerceCustomer(ctx context.Context) (*dtocustomer.CustomerResult, error)
 	CommerceCart(ctx context.Context) (*dto.DecoratedCart, error)
 	CommerceCartValidator(ctx context.Context) (*validation.Result, error)
+	CommerceCartQtyRestriction(ctx context.Context, marketplaceCode string, variantCode *string, deliveryCode string) (*validation.RestrictionResult, error)
 	CommerceCheckoutActivePlaceOrder(ctx context.Context) (bool, error)
 	CommerceCheckoutCurrentContext(ctx context.Context) (*dto2.PlaceOrderContext, error)
 	CommerceCategoryTree(ctx context.Context, activeCategoryCode string) (domain2.Tree, error)
@@ -2410,6 +2419,34 @@ func (e *executableSchema) Complexity(typeName, field string, childComplexity in
 		}
 
 		return e.complexity.CommerceCartPricedTotalItem.Code(childComplexity), true
+
+	case "Commerce_Cart_QtyRestrictionResult.isRestricted":
+		if e.complexity.CommerceCartQtyRestrictionResult.IsRestricted == nil {
+			break
+		}
+
+		return e.complexity.CommerceCartQtyRestrictionResult.IsRestricted(childComplexity), true
+
+	case "Commerce_Cart_QtyRestrictionResult.maxAllowed":
+		if e.complexity.CommerceCartQtyRestrictionResult.MaxAllowed == nil {
+			break
+		}
+
+		return e.complexity.CommerceCartQtyRestrictionResult.MaxAllowed(childComplexity), true
+
+	case "Commerce_Cart_QtyRestrictionResult.remainingDifference":
+		if e.complexity.CommerceCartQtyRestrictionResult.RemainingDifference == nil {
+			break
+		}
+
+		return e.complexity.CommerceCartQtyRestrictionResult.RemainingDifference(childComplexity), true
+
+	case "Commerce_Cart_QtyRestrictionResult.restrictorName":
+		if e.complexity.CommerceCartQtyRestrictionResult.RestrictorName == nil {
+			break
+		}
+
+		return e.complexity.CommerceCartQtyRestrictionResult.RestrictorName(childComplexity), true
 
 	case "Commerce_Cart_SelectedPaymentResult.processed":
 		if e.complexity.CommerceCartSelectedPaymentResult.Processed == nil {
@@ -4139,6 +4176,18 @@ func (e *executableSchema) Complexity(typeName, field string, childComplexity in
 
 		return e.complexity.Query.CommerceCart(childComplexity), true
 
+	case "Query.Commerce_Cart_QtyRestriction":
+		if e.complexity.Query.CommerceCartQtyRestriction == nil {
+			break
+		}
+
+		args, err := ec.field_Query_Commerce_Cart_QtyRestriction_args(context.TODO(), rawArgs)
+		if err != nil {
+			return 0, false
+		}
+
+		return e.complexity.Query.CommerceCartQtyRestriction(childComplexity, args["marketplaceCode"].(string), args["variantCode"].(*string), args["deliveryCode"].(string)), true
+
 	case "Query.Commerce_Cart_Validator":
 		if e.complexity.Query.CommerceCartValidator == nil {
 			break
@@ -4594,6 +4643,14 @@ type Commerce_Cart_ItemValidationError {
     errorMessageKey: String!
 }
 
+
+type Commerce_Cart_QtyRestrictionResult {
+    isRestricted:        Boolean!
+    maxAllowed:          Int!
+    remainingDifference: Int!
+    restrictorName:      String!
+}
+
 type Commerce_Cart_PlacedOrderInfo {
     orderNumber:    String!
     deliveryCode:   String!
@@ -4722,6 +4779,8 @@ input Commerce_Cart_DeliveryShippingOption {
 extend type Query {
     Commerce_Cart: Commerce_DecoratedCart!
     Commerce_Cart_Validator: Commerce_Cart_ValidationResult!
+    "Commerce_Cart_QtyRestriction returns if the product is restricted in terms of the allowed quantity for the current cart and the given delivery"
+    Commerce_Cart_QtyRestriction(marketplaceCode: String!, variantCode: String, deliveryCode: String!): Commerce_Cart_QtyRestrictionResult!
 }
 
 extend type Mutation {
@@ -5847,6 +5906,36 @@ func (ec *executionContext) field_Mutation_Commerce_UpdateItemQty_args(ctx conte
 		}
 	}
 	args["qty"] = arg2
+	return args, nil
+}
+
+func (ec *executionContext) field_Query_Commerce_Cart_QtyRestriction_args(ctx context.Context, rawArgs map[string]interface{}) (map[string]interface{}, error) {
+	var err error
+	args := map[string]interface{}{}
+	var arg0 string
+	if tmp, ok := rawArgs["marketplaceCode"]; ok {
+		arg0, err = ec.unmarshalNString2string(ctx, tmp)
+		if err != nil {
+			return nil, err
+		}
+	}
+	args["marketplaceCode"] = arg0
+	var arg1 *string
+	if tmp, ok := rawArgs["variantCode"]; ok {
+		arg1, err = ec.unmarshalOString2ᚖstring(ctx, tmp)
+		if err != nil {
+			return nil, err
+		}
+	}
+	args["variantCode"] = arg1
+	var arg2 string
+	if tmp, ok := rawArgs["deliveryCode"]; ok {
+		arg2, err = ec.unmarshalNString2string(ctx, tmp)
+		if err != nil {
+			return nil, err
+		}
+	}
+	args["deliveryCode"] = arg2
 	return args, nil
 }
 
@@ -11648,6 +11737,114 @@ func (ec *executionContext) _Commerce_Cart_PricedTotalItem_code(ctx context.Cont
 	resTmp := ec.FieldMiddleware(ctx, obj, func(rctx context.Context) (interface{}, error) {
 		ctx = rctx // use context from middleware stack in children
 		return obj.Code, nil
+	})
+	if resTmp == nil {
+		if !ec.HasError(rctx) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(string)
+	rctx.Result = res
+	ctx = ec.Tracer.StartFieldChildExecution(ctx)
+	return ec.marshalNString2string(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) _Commerce_Cart_QtyRestrictionResult_isRestricted(ctx context.Context, field graphql.CollectedField, obj *validation.RestrictionResult) graphql.Marshaler {
+	ctx = ec.Tracer.StartFieldExecution(ctx, field)
+	defer func() { ec.Tracer.EndFieldExecution(ctx) }()
+	rctx := &graphql.ResolverContext{
+		Object:   "Commerce_Cart_QtyRestrictionResult",
+		Field:    field,
+		Args:     nil,
+		IsMethod: false,
+	}
+	ctx = graphql.WithResolverContext(ctx, rctx)
+	ctx = ec.Tracer.StartFieldResolverExecution(ctx, rctx)
+	resTmp := ec.FieldMiddleware(ctx, obj, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.IsRestricted, nil
+	})
+	if resTmp == nil {
+		if !ec.HasError(rctx) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(bool)
+	rctx.Result = res
+	ctx = ec.Tracer.StartFieldChildExecution(ctx)
+	return ec.marshalNBoolean2bool(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) _Commerce_Cart_QtyRestrictionResult_maxAllowed(ctx context.Context, field graphql.CollectedField, obj *validation.RestrictionResult) graphql.Marshaler {
+	ctx = ec.Tracer.StartFieldExecution(ctx, field)
+	defer func() { ec.Tracer.EndFieldExecution(ctx) }()
+	rctx := &graphql.ResolverContext{
+		Object:   "Commerce_Cart_QtyRestrictionResult",
+		Field:    field,
+		Args:     nil,
+		IsMethod: false,
+	}
+	ctx = graphql.WithResolverContext(ctx, rctx)
+	ctx = ec.Tracer.StartFieldResolverExecution(ctx, rctx)
+	resTmp := ec.FieldMiddleware(ctx, obj, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.MaxAllowed, nil
+	})
+	if resTmp == nil {
+		if !ec.HasError(rctx) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(int)
+	rctx.Result = res
+	ctx = ec.Tracer.StartFieldChildExecution(ctx)
+	return ec.marshalNInt2int(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) _Commerce_Cart_QtyRestrictionResult_remainingDifference(ctx context.Context, field graphql.CollectedField, obj *validation.RestrictionResult) graphql.Marshaler {
+	ctx = ec.Tracer.StartFieldExecution(ctx, field)
+	defer func() { ec.Tracer.EndFieldExecution(ctx) }()
+	rctx := &graphql.ResolverContext{
+		Object:   "Commerce_Cart_QtyRestrictionResult",
+		Field:    field,
+		Args:     nil,
+		IsMethod: false,
+	}
+	ctx = graphql.WithResolverContext(ctx, rctx)
+	ctx = ec.Tracer.StartFieldResolverExecution(ctx, rctx)
+	resTmp := ec.FieldMiddleware(ctx, obj, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.RemainingDifference, nil
+	})
+	if resTmp == nil {
+		if !ec.HasError(rctx) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(int)
+	rctx.Result = res
+	ctx = ec.Tracer.StartFieldChildExecution(ctx)
+	return ec.marshalNInt2int(ctx, field.Selections, res)
+}
+
+func (ec *executionContext) _Commerce_Cart_QtyRestrictionResult_restrictorName(ctx context.Context, field graphql.CollectedField, obj *validation.RestrictionResult) graphql.Marshaler {
+	ctx = ec.Tracer.StartFieldExecution(ctx, field)
+	defer func() { ec.Tracer.EndFieldExecution(ctx) }()
+	rctx := &graphql.ResolverContext{
+		Object:   "Commerce_Cart_QtyRestrictionResult",
+		Field:    field,
+		Args:     nil,
+		IsMethod: false,
+	}
+	ctx = graphql.WithResolverContext(ctx, rctx)
+	ctx = ec.Tracer.StartFieldResolverExecution(ctx, rctx)
+	resTmp := ec.FieldMiddleware(ctx, obj, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return obj.RestrictorName, nil
 	})
 	if resTmp == nil {
 		if !ec.HasError(rctx) {
@@ -17949,6 +18146,40 @@ func (ec *executionContext) _Query_Commerce_Cart_Validator(ctx context.Context, 
 	return ec.marshalNCommerce_Cart_ValidationResult2ᚖflamingoᚗmeᚋflamingoᚑcommerceᚋv3ᚋcartᚋdomainᚋvalidationᚐResult(ctx, field.Selections, res)
 }
 
+func (ec *executionContext) _Query_Commerce_Cart_QtyRestriction(ctx context.Context, field graphql.CollectedField) graphql.Marshaler {
+	ctx = ec.Tracer.StartFieldExecution(ctx, field)
+	defer func() { ec.Tracer.EndFieldExecution(ctx) }()
+	rctx := &graphql.ResolverContext{
+		Object:   "Query",
+		Field:    field,
+		Args:     nil,
+		IsMethod: true,
+	}
+	ctx = graphql.WithResolverContext(ctx, rctx)
+	rawArgs := field.ArgumentMap(ec.Variables)
+	args, err := ec.field_Query_Commerce_Cart_QtyRestriction_args(ctx, rawArgs)
+	if err != nil {
+		ec.Error(ctx, err)
+		return graphql.Null
+	}
+	rctx.Args = args
+	ctx = ec.Tracer.StartFieldResolverExecution(ctx, rctx)
+	resTmp := ec.FieldMiddleware(ctx, nil, func(rctx context.Context) (interface{}, error) {
+		ctx = rctx // use context from middleware stack in children
+		return ec.resolvers.Query().CommerceCartQtyRestriction(rctx, args["marketplaceCode"].(string), args["variantCode"].(*string), args["deliveryCode"].(string))
+	})
+	if resTmp == nil {
+		if !ec.HasError(rctx) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	res := resTmp.(*validation.RestrictionResult)
+	rctx.Result = res
+	ctx = ec.Tracer.StartFieldChildExecution(ctx)
+	return ec.marshalNCommerce_Cart_QtyRestrictionResult2ᚖflamingoᚗmeᚋflamingoᚑcommerceᚋv3ᚋcartᚋdomainᚋvalidationᚐRestrictionResult(ctx, field.Selections, res)
+}
+
 func (ec *executionContext) _Query_Commerce_Checkout_ActivePlaceOrder(ctx context.Context, field graphql.CollectedField) graphql.Marshaler {
 	ctx = ec.Tracer.StartFieldExecution(ctx, field)
 	defer func() { ec.Tracer.EndFieldExecution(ctx) }()
@@ -20946,6 +21177,48 @@ func (ec *executionContext) _Commerce_Cart_PricedTotalItem(ctx context.Context, 
 	return out
 }
 
+var commerce_Cart_QtyRestrictionResultImplementors = []string{"Commerce_Cart_QtyRestrictionResult"}
+
+func (ec *executionContext) _Commerce_Cart_QtyRestrictionResult(ctx context.Context, sel ast.SelectionSet, obj *validation.RestrictionResult) graphql.Marshaler {
+	fields := graphql.CollectFields(ec.RequestContext, sel, commerce_Cart_QtyRestrictionResultImplementors)
+
+	out := graphql.NewFieldSet(fields)
+	var invalids uint32
+	for i, field := range fields {
+		switch field.Name {
+		case "__typename":
+			out.Values[i] = graphql.MarshalString("Commerce_Cart_QtyRestrictionResult")
+		case "isRestricted":
+			out.Values[i] = ec._Commerce_Cart_QtyRestrictionResult_isRestricted(ctx, field, obj)
+			if out.Values[i] == graphql.Null {
+				invalids++
+			}
+		case "maxAllowed":
+			out.Values[i] = ec._Commerce_Cart_QtyRestrictionResult_maxAllowed(ctx, field, obj)
+			if out.Values[i] == graphql.Null {
+				invalids++
+			}
+		case "remainingDifference":
+			out.Values[i] = ec._Commerce_Cart_QtyRestrictionResult_remainingDifference(ctx, field, obj)
+			if out.Values[i] == graphql.Null {
+				invalids++
+			}
+		case "restrictorName":
+			out.Values[i] = ec._Commerce_Cart_QtyRestrictionResult_restrictorName(ctx, field, obj)
+			if out.Values[i] == graphql.Null {
+				invalids++
+			}
+		default:
+			panic("unknown field " + strconv.Quote(field.Name))
+		}
+	}
+	out.Dispatch()
+	if invalids > 0 {
+		return graphql.Null
+	}
+	return out
+}
+
 var commerce_Cart_SelectedPaymentResultImplementors = []string{"Commerce_Cart_SelectedPaymentResult"}
 
 func (ec *executionContext) _Commerce_Cart_SelectedPaymentResult(ctx context.Context, sel ast.SelectionSet, obj *dto.SelectedPaymentResult) graphql.Marshaler {
@@ -23233,6 +23506,20 @@ func (ec *executionContext) _Query(ctx context.Context, sel ast.SelectionSet) gr
 				}
 				return res
 			})
+		case "Commerce_Cart_QtyRestriction":
+			field := field
+			out.Concurrently(i, func() (res graphql.Marshaler) {
+				defer func() {
+					if r := recover(); r != nil {
+						ec.Error(ctx, ec.Recover(ctx, r))
+					}
+				}()
+				res = ec._Query_Commerce_Cart_QtyRestriction(ctx, field)
+				if res == graphql.Null {
+					atomic.AddUint32(&invalids, 1)
+				}
+				return res
+			})
 		case "Commerce_Checkout_ActivePlaceOrder":
 			field := field
 			out.Concurrently(i, func() (res graphql.Marshaler) {
@@ -23737,6 +24024,20 @@ func (ec *executionContext) marshalNCommerce_Cart_PricedShippingItem2flamingoᚗ
 
 func (ec *executionContext) marshalNCommerce_Cart_PricedTotalItem2flamingoᚗmeᚋflamingoᚑcommerceᚋv3ᚋcartᚋinterfacesᚋgraphqlᚋdtoᚐPricedTotalItem(ctx context.Context, sel ast.SelectionSet, v dto.PricedTotalItem) graphql.Marshaler {
 	return ec._Commerce_Cart_PricedTotalItem(ctx, sel, &v)
+}
+
+func (ec *executionContext) marshalNCommerce_Cart_QtyRestrictionResult2flamingoᚗmeᚋflamingoᚑcommerceᚋv3ᚋcartᚋdomainᚋvalidationᚐRestrictionResult(ctx context.Context, sel ast.SelectionSet, v validation.RestrictionResult) graphql.Marshaler {
+	return ec._Commerce_Cart_QtyRestrictionResult(ctx, sel, &v)
+}
+
+func (ec *executionContext) marshalNCommerce_Cart_QtyRestrictionResult2ᚖflamingoᚗmeᚋflamingoᚑcommerceᚋv3ᚋcartᚋdomainᚋvalidationᚐRestrictionResult(ctx context.Context, sel ast.SelectionSet, v *validation.RestrictionResult) graphql.Marshaler {
+	if v == nil {
+		if !ec.HasError(graphql.GetResolverContext(ctx)) {
+			ec.Errorf(ctx, "must not be null")
+		}
+		return graphql.Null
+	}
+	return ec._Commerce_Cart_QtyRestrictionResult(ctx, sel, v)
 }
 
 func (ec *executionContext) marshalNCommerce_Cart_SelectedPaymentResult2flamingoᚗmeᚋflamingoᚑcommerceᚋv3ᚋcartᚋinterfacesᚋgraphqlᚋdtoᚐSelectedPaymentResult(ctx context.Context, sel ast.SelectionSet, v dto.SelectedPaymentResult) graphql.Marshaler {